+++ web/WEB-INF/classes/src/org/oscarehr/PMmodule/service/ProgramManager.java
@@ -135,7 +135,7 @@ public class ProgramManager {
         return programDao.getAllPrograms();
     }
 
-    public List<Program> getAllPrograms(String programStatus, String type, long facilityId) {
+    public List<Program> getAllPrograms(String programStatus, String type, int facilityId) {
         return programDao.getAllPrograms(programStatus, type, facilityId);
     }
 
@@ -349,7 +349,7 @@ public class ProgramManager {
         return programDomain;
     }
 
-    public List<Program> getProgramDomainInFacility(String providerNo, Long facilityId) {
+    public List<Program> getProgramDomainInFacility(String providerNo, Integer facilityId) {
     	List<Program> programs = getProgramDomain(providerNo);
     	List<Program> results = new ArrayList<Program>();
     	if(facilityId==null)
